Sprint-3 POS application completed.

+++ src/main/java/com/ust/pos/dto/PriceDto.java
@@ -3,11 +3,13 @@ package com.ust.pos.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class PriceDto extends CommonDto {
 
     private String product;
-    private Double priceAmount;
+    private BigDecimal priceAmount;
     private String priceType;
 }
