+++ src/main/java/com/ust/pos/model/Price.java
@@ -4,12 +4,14 @@ import jakarta.persistence.Entity;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Entity
 public class Price extends CommonFields {
 
     private String product;
-    private Double priceAmount;
+    private BigDecimal priceAmount;
     private String priceType;
 }
